+++ bankfeeds/RELEASES.md
@@ -667,3 +667,13 @@ Based on:
 - [typescript v0.36.0] bankfeeds
 ### Releases
 - [NPM v0.36.0] https://www.npmjs.com/package/@codat/bank-feeds/v/0.36.0 - bankfeeds
+
+## 2023-09-01 00:04:34
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.77.0 (2.91.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.0] bankfeeds
+### Releases
+- [NPM v0.37.0] https://www.npmjs.com/package/@codat/bank-feeds/v/0.37.0 - bankfeeds